+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All notable changes to Sourcegraph Server and Data Center are documented in this
 
 * Fixed an issue where langservers that were not enabled would display the "Restart" button in the Code Intelligence management panel.
 * Fixed an issue where the "Update" button in the Code Intelligence management panel would be displayed inconsistently.
+* Fixed an issue where toggling a dynamic search scope would not also remove `@rev` (if specified)
 
 ### Added